Spatial Access Trees (SAT)

SpatialAccessTrees.Sat — Type

struct Sat{DT<:SemiMetric,DBT<:AbstractDatabase} <: AbstractSearchIndex
    dist::DT
    db::DBT
    root::UInt32
    # parents::Vector{UInt32}
    children::Vector{Union{Nothing,Vector{UInt32}}}
    cov::Vector{Float32} # leafs: ``- d(parent, leaf)``, internal: ``\max \{d(parent, u) | u \in children(parent)\}``
end

Spatial Access Tree data structure. Please see Sat function for high level constructors

SimilaritySearch.index! — Function

index!(
    sat::Sat;
    sortsat::AbstractSortSat=ProximalSortSat(),
    minleaf::Int=log2(ceil(database(sat)))
)

Performs the indexing of the referenced dataset in the spatial access tree.

Arguments

sat: The metric data structure

Keyword arguments

sortsat: The strategy to create the spatial access tree, it heavily depends on the order of elements while it is build. It accepts:

RandomSortSat(): children are randomized.

ProximalSortSat(): classical approach, near elements are put first.

DistalSortSat(): recent approach, distant elements are put first.

minleaf: Minimum number of children to perform a spatial access separation (half space partitioning)

SAT orders

Missing docstring.

Missing docstring for ProximalSortSat. Check Documenter's build log for details.

Missing docstring.

Missing docstring for DistalSortSat. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RandomSortSat. Check Documenter's build log for details.

Approximate similarity search

SpatialAccessTrees.PruningSat — Type

PruningSat(sat::Sat; factor=0.9)

Creates an approximate similarity sarch index based on sat and aggressive pruning; adapted from the paper A probabilistic spell for the curse of dimensionality (Chavez and Navarro, 2001). It supports auto-tuning via optimize!.

SpatialAccessTrees.BeamSearchSat — Type

BeamSearchSat(sat::Sat; bsize=8, Δ=1f0)

Creates an approximate similarity sarch index based on sat and aggressive pruning; adapted from the paper A probabilistic spell for the curse of dimensionality (Chavez and Navarro, 2001). It supports auto-tuning via optimize!.

SpatialAccessTrees.BeamSearchMultiSat — Type

BeamSearchMultiSat(sat::Vector{SatType}; bsize=8, Δ=1f0) where {SatType<:Sat}

Applies beam search on a multi SAT array. See optimize! to tune the index for some objective.

Missing docstring.

Missing docstring for optimize!. Check Documenter's build log for details.